Install manifest to both lib directories
According to documentation, the manifest file can be put either in /usr/lib64/mozilla/native-messaging-hosts or /usr/lib/mozilla/native-messaging-hosts.

The Makefile made the decision where to put it based on the machine's architecture.
On Ubuntu however, the manifest in /usr/lib64 is not recognized, even if arch in x86_64.

I don't want to maintain a list of distributions where they expect the manifest, so I'm now copying it to both locations.