Wait a moment after first connection to KGet fails
DBus will start KGet if it's not running on sending the download request. But the download request will be "forgotten".

There already was a second connection attempt implemented but it came before KGet would be started properly.

This introduces a small wait time in case the first connection failed, assuming that KGet is now in the process of starting up.